Export Alert to viewpage
